split DVD* functions to dvd.h
