+++ include/dvd/dvd.h
@@ -17,7 +17,27 @@ typedef enum DVDState {
     DVD_STATE_RETRY         = 0xb,
     DVD_STATE_FATAL_ERROR   = 0xffffffff,
 } DVDState;
+}
+
+class DVDFileInfo;
+extern "C" {
+    s32 DVDOpen(const char*, u8[48]);
+    s32 DVDClose(u8[48]);
+    void DVDReadPrio(void);
+    void DVDGetCurrentDiskID(void);
+    s32 DVDFastOpen(long, u8[48]);
+    int DVDGetCommandBlockStatus(u8[48]);
+    s32 DVDReadAsyncPrio(u8[48], void*, long, long, void(*)(long,DVDFileInfo*), long);
+    void DVDConvertPathToEntrynum(void);
+    DVDState DVDGetDriveStatus(void);
+    s32 DVDCheckDisk(void);
+
+    void DVDChangeDir(void);
+    void DVDCloseDir(void);
+    void DVDOpenDir(void);
+    void DVDReadDir(void);
 
 }
+
 
 #endif
